# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "VK TPA Website API Documentation",
        "title": "TPA-Website",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/hotel-transaction": {
            "post": {
                "description": "Add new hotel transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hotel Transaction"
                ],
                "summary": "Add new hotel transaction",
                "parameters": [
                    {
                        "description": "Hotel Transaction Body",
                        "name": "hotelTransaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hotel transaction created successfully!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/hotel-transaction/hotel-cart": {
            "post": {
                "description": "Add new hotel cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hotel Cart"
                ],
                "summary": "Add new hotel cart",
                "parameters": [
                    {
                        "description": "Hotel Cart Body",
                        "name": "hotelCart",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hotel cart created successfully!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/hotel-transaction/user/history/{id}": {
            "get": {
                "description": "Get user history hotel transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hotel Transaction"
                ],
                "summary": "Get user history hotel transaction",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hotel transaction created successfully!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/hotel-transaction/user/ongoing/{id}": {
            "get": {
                "description": "Get user ongoing hotel transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hotel Transaction"
                ],
                "summary": "Get user ongoing hotel transaction",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hotel transaction created successfully!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
}